inheritance work
Program one : app.java : The IndoorActivity and outdoor activity class represents indoor activities. It inherits from the ActivityBase class and provides additional information specific to indoor activities. An indoor activity typically takes place indoors, such as in a house or a facility, and is designed for a specific number of participants.

program two : que2.java : The Weather class serves as the base class for representing weather conditions. It implements the WeatherCondition interface, providing a common structure and behavior for describing various weather conditions.